some fixes

+++ modules/harvest/src/HarvestService.php
@@ -144,18 +144,21 @@ class HarvestService implements ContainerInjectionInterface {
    *   Whether this happened successfully.
    */
   public function deregisterHarvest(string $plan_id) {
-    // Remove all the support tables for this plan id.
-    foreach ([
-      'harvest_' . $plan_id . '_items',
-      'harvest_' . $plan_id . '_hashes',
-      'harvest_' . $plan_id . '_runs',
-    ] as $table_name) {
-      /** @var \Drupal\common\Storage\DatabaseTableInterface $store */
-      $store = $this->storeFactory->getInstance($table_name);
-      $store->destruct();
+    if (in_array($plan_id, $this->getAllHarvestIds())) {
+      // Remove all the support tables for this plan id.
+      foreach ([
+        'harvest_' . $plan_id . '_items',
+        'harvest_' . $plan_id . '_hashes',
+        'harvest_' . $plan_id . '_runs',
+      ] as $table_name) {
+        /** @var \Drupal\common\Storage\DatabaseTableInterface $store */
+        $store = $this->storeFactory->getInstance($table_name);
+        $store->destruct();
+      }
+      // Remove the plan id from the harvest_plans table.
+      return $this->harvestPlanRepository->remove($plan_id);
     }
-    // Remove the plan id from the harvest_plans table.
-    return $this->harvestPlanRepository->remove($plan_id);
+    return FALSE;
   }
 
   /**
@@ -165,8 +168,8 @@ class HarvestService implements ContainerInjectionInterface {
    */
   public function revertHarvest($id) {
     $run_store = $this->storeFactory->getInstance("harvest_{$id}_runs");
-    if (!method_exists($run_store, "destruct")) {
-      throw new \Exception("Storage of class " . get_class($run_store) . " does not implement destruct method.");
+    if (!method_exists($run_store, 'destruct')) {
+      throw new \Exception('Storage of class ' . get_class($run_store) . ' does not implement destruct method.');
     }
     $run_store->destruct();
     $harvester = $this->getHarvester($id);
@@ -181,7 +184,7 @@ class HarvestService implements ContainerInjectionInterface {
 
     $result = $harvester->harvest();
     if (is_null($result['status']['extracted_items_ids'] ?? NULL)) {
-      throw new \Exception("No items found to extract, review your harvest plan.");
+      throw new \Exception('No items found to extract, review your harvest plan.');
     }
     $result['status']['orphan_ids'] = $this->getOrphanIdsFromResult($id, $result['status']['extracted_items_ids']);
     $this->processOrphanIds($result['status']['orphan_ids']);
